data_pipeline_v3/requirements.txt
pandas
numpy
yfinance
pandas_datareader
pyyaml
scikit-learn
pyarrow
ta
transformers
torch
newsapi-python
requests